# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CMMS API — Beit Jala Pharmaceutical Co.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /machines/{id}/corrective-maintenance:
    get:
      operationId: getMachineCorrectiveMaintenance
      tags: [corrective-maintenance]
      summary: Get active Corrective Maintenance log for a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Corrective Maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorrectiveMaintenanceRecordDetail"
  /machines/{id}/corrective-maintenance/history:
    get:
      operationId: getMachineCorrectiveMaintenanceHistory
      tags: [corrective-maintenance]
      summary: List Corrective Maintenance logs for a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Corrective Maintenance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorrectiveMaintenanceRecordDetail"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    AuthUser:
      type: object
      required: [id, username, roleId, roleName, permissions, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        roleId:
          type: integer
        roleName:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    PerformingStaffInput:
      type: object
      properties:
        no:
          type: string
        name:
          type: string
        signature:
          type: string
...
    CorrectiveMaintenanceEvent:
      type: object
      required: [id, recordId, requestId, machineId, requestReportNumber, rowNumber]
      properties:
        id:
          type: integer
        recordId:
          type: integer
        requestId:
          type: integer
        machineId:
          type: integer
        requestReportNumber:
          type: string
        rowNumber:
          type: integer
        preliminaryCheckResults:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        maintenanceTechnicianSignature:
          type: ["string", "null"]
        concernedSectionSupervisorSignature:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        remarksRecommendations:
          type: ["string", "null"]
        performingStaff:
          type: array
          items:
            $ref: "#/components/schemas/PerformingStaffInput"
        receiverName:
          type: ["string", "null"]
        receiverSignature:
          type: ["string", "null"]
        handoverDate:
          type: ["string", "null"]
        engineeringSignature:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
    CorrectiveMaintenanceRecordDetail:
      type: object
      required: [id, machineId, sequenceNumber, documentNumber, pageCount, machineName, machineNumber, status, events]
      properties:
        id:
          type: integer
        machineId:
          type: integer
        sequenceNumber:
          type: integer
        documentNumber:
          type: string
        executionDate:
          type: ["string", "null"]
        pageCount:
          type: string
        machineName:
          type: string
        machineNumber:
          type: string
        machineLocation:
          type: ["string", "null"]
        startupDate:
          type: ["string", "null"]
        maxRows:
          type: integer
        status:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/CorrectiveMaintenanceEvent"
...